updateGeral

+++ src/main/java/com/example/demo/domain/local/RegisterLocalDTO.java
@@ -3,5 +3,5 @@ package com.example.demo.domain.local;
 
 import com.example.demo.domain.user.User;
 
-public record RegisterLocalDTO(String descricao, float price, User userId, String endereco) {
+public record RegisterLocalDTO(String descricao, Integer price, User userId, String endereco) {
 }
